+++ packages/dara-core/dara/core/internal/routing.py
@@ -23,7 +23,16 @@ from typing import Any, Callable, Dict, List, Mapping, Optional
 
 import anyio
 import pandas
-from fastapi import APIRouter, Body, Depends, File, Form, HTTPException, Response, UploadFile
+from fastapi import (
+    APIRouter,
+    Body,
+    Depends,
+    File,
+    Form,
+    HTTPException,
+    Response,
+    UploadFile,
+)
 from fastapi.responses import StreamingResponse
 from pandas import DataFrame
 from pydantic import BaseModel
@@ -56,7 +65,7 @@ from dara.core.internal.registry_lookup import RegistryLookup
 from dara.core.internal.settings import get_settings
 from dara.core.internal.tasks import TaskManager, TaskManagerError
 from dara.core.internal.utils import get_cache_scope
-from dara.core.internal.websocket import WS_CHANNEL, WebsocketManager, ws_handler
+from dara.core.internal.websocket import WS_CHANNEL, ws_handler
 from dara.core.logging import dev_logger
 from dara.core.persistence import BackendStoreEntry
 from dara.core.visual.dynamic_component import CURRENT_COMPONENT_ID, PyComponentDef