+++ packages/dara-core/dara/core/persistence.py
@@ -1,9 +1,9 @@
 import abc
-import asyncio
-import uuid
 from typing import TYPE_CHECKING, Any
 
 from pydantic import BaseModel, Field
+
+from dara.core.internal.utils import run_user_handler
 
 if TYPE_CHECKING:
     from dara.core.interactivity.plain_variable import Variable
@@ -21,21 +21,18 @@ class PersistenceBackend(BaseModel, abc.ABC):
         """
         Persist a value
         """
-        pass
 
     @abc.abstractmethod
     def read(self, key: str) -> Any:
         """
         Read a value
         """
-        pass
 
     @abc.abstractmethod
     def delete(self, key: str):
         """
         Delete a value
         """
-        pass
 
 
 class InMemoryBackend(PersistenceBackend):
@@ -64,11 +61,10 @@ class PersistenceStore(BaseModel, abc.ABC):
     # TODO: if need be this can also hold a reference to js impl
 
     @abc.abstractmethod
-    def init(self, variable: 'Variable'):
+    async def init(self, variable: 'Variable'):
         """
         Initialize the store when connecting to a variable
         """
-        pass
 
     def dict(self, *args, **kwargs):
         parent_dict = super().dict(*args, **kwargs)
@@ -78,8 +74,7 @@ class PersistenceStore(BaseModel, abc.ABC):
 
 class BackendStore(PersistenceStore):
     """
-    Persistence store implementation that uses a backend to store data server-side
-
+    Persistence store implementation that uses a backend implementation to store data server-side
     """
 
     uid: str
@@ -111,38 +106,43 @@ class BackendStore(PersistenceStore):
         ws_mgr: WebsocketManager = utils_registry.get('WebsocketManager')
         await ws_mgr.broadcast({'store_uid': self.uid, 'value': value})
 
-    def init(self, variable: 'Variable'):
+    async def init(self, variable: 'Variable'):
         """
         Write the default value to the store if it's not set
         """
-        if self.read() is None:
-            self.write(variable.default, notify=False)
+        if await self.read() is None:
+            await self.write(variable.default, notify=False)
 
-    def write(self, value: Any, notify=True):
+    async def write(self, value: Any, notify=True):
         """
         Persist a value to the store
+
+        :param value: value to write
+        :param notify: whether to broadcast the new value to clients
         """
         if notify:
             # Schedule notification on write
-            asyncio.create_task(self._notify(value))
+            await self._notify(value)
 
         # TODO: in the future key can be self.uid + user_id if scope='user'
-        return self.backend.write(self.uid, value)
+        return await run_user_handler(self.backend.write, (self.uid, value))
 
-    def read(self):
+    async def read(self):
         """
         Read a value from the store
         """
-        return self.backend.read(self.uid)
+        return await run_user_handler(self.backend.read, (self.uid,))
 
-    def delete(self, notify=True):
+    async def delete(self, notify=True):
         """
         Delete the persisted value from the store
+
+        :param notify: whether to broadcast that the value was deleted to clients
         """
         if notify:
             # Schedule notification on delete
-            asyncio.create_task(self._notify(None))
-        return self.backend.delete(self.uid)
+            await self._notify(None)
+        return await run_user_handler(self.backend.delete, (self.uid,))
 
 
 class BackendStoreEntry(BaseModel):